add TVL related fields to Sushi
